Add tools to convert for medium syndication (#3)
For reasons that defy my imagining, medium still does not support
latex. Also, it has no API to speak of as far as managing posts. These
scripts will convert blocks of latex to images and inline latex to
unicode characters that are, hopefully, roughly equivalent. I "vibe"
coded those scripts so they are a bit of a mess.